Delete duplicate test_framework_utils
The test_framework_utils.py tests were mostly duplicated in test_framework_helpers.py already.  This copies some bits over and then deletes the file with duplicate tests.
